+++ contracts/p1/BackingManager.sol
@@ -113,7 +113,6 @@ contract BackingManagerP1 is TradingP1, IBackingManager {
     function rebalance(TradeKind kind) external nonReentrant notTradingPausedOrFrozen {
         // == Refresh ==
         assetRegistry.refresh();
-        furnace.melt();
 
         // DoS prevention: unless caller is self, require 1 empty block between like-kind auctions
         require(
@@ -184,7 +183,6 @@ contract BackingManagerP1 is TradingP1, IBackingManager {
         require(ArrayLib.allUnique(erc20s), "duplicate tokens");
 
         assetRegistry.refresh();
-        furnace.melt();
 
         BasketRange memory basketsHeld = basketHandler.basketsHeldBy(address(this));
 
@@ -212,10 +210,12 @@ contract BackingManagerP1 is TradingP1, IBackingManager {
          *      RToken traders according to the distribution totals.
          */
 
-        // Forward any RSR held to StRSR pool; RSR should never be sold for RToken yield
+        // Forward any RSR held to StRSR pool and payout rewards
+        // RSR should never be sold for RToken yield
         if (rsr.balanceOf(address(this)) > 0) {
             // For CEI, this is an interaction "within our system" even though RSR is already live
             IERC20(address(rsr)).safeTransfer(address(stRSR), rsr.balanceOf(address(this)));
+            stRSR.payoutRewards();
         }
 
         // Mint revenue RToken