+++ contracts/p1/RevenueTrader.sol
@@ -134,10 +134,8 @@ contract RevenueTraderP1 is TradingP1, IRevenueTrader {
         IAsset assetToBuy = assetRegistry.toAsset(tokenToBuy);
 
         // Refresh everything if RToken is involved
-        if (involvesRToken) {
-            assetRegistry.refresh();
-            furnace.melt();
-        } else {
+        if (involvesRToken) assetRegistry.refresh();
+        else {
             // Otherwise: refresh just the needed assets and nothing more
             for (uint256 i = 0; i < len; ++i) {
                 assetRegistry.toAsset(erc20s[i]).refresh();
@@ -190,6 +188,9 @@ contract RevenueTraderP1 is TradingP1, IRevenueTrader {
         uint256 bal = tokenToBuy.balanceOf(address(this));
         tokenToBuy.safeApprove(address(distributor), 0);
         tokenToBuy.safeApprove(address(distributor), bal);
+
+        // do not need to use AllowanceLib.safeApproveFallbackToCustom here because
+        // tokenToBuy can be assumed to be either RSR or the RToken
         distributor.distribute(tokenToBuy, bal);
     }