+++ test/Revenues.test.ts
@@ -13,6 +13,7 @@ import {
   CollateralStatus,
   TradeKind,
   MAX_UINT192,
+  ONE_PERIOD,
 } from '../common/constants'
 import { expectEvents } from '../common/events'
 import { bn, divCeil, fp, near } from '../common/numbers'
@@ -555,7 +556,7 @@ describe(`Revenues - P${IMPLEMENTATION}`, () => {
         )
       })
 
-      it('Should forward RSR revenue directly to StRSR', async () => {
+      it('Should forward RSR revenue directly to StRSR and call payoutRewards()', async () => {
         const amount = bn('2000e18')
         await rsr.connect(owner).mint(backingManager.address, amount)
         expect(await rsr.balanceOf(backingManager.address)).to.equal(amount)
@@ -563,7 +564,23 @@ describe(`Revenues - P${IMPLEMENTATION}`, () => {
         expect(await rsr.balanceOf(rsrTrader.address)).to.equal(0)
         expect(await rsr.balanceOf(rTokenTrader.address)).to.equal(0)
 
-        await expect(backingManager.forwardRevenue([rsr.address])).to.emit(rsr, 'Transfer')
+        // Advance to the end of noop period
+        await advanceTime(Number(ONE_PERIOD))
+
+        await expectEvents(backingManager.forwardRevenue([rsr.address]), [
+          {
+            contract: rsr,
+            name: 'Transfer',
+            args: [backingManager.address, stRSR.address, amount],
+            emitted: true,
+          },
+          {
+            contract: stRSR,
+            name: 'RewardsPaid',
+            emitted: true,
+          },
+        ])
+
         expect(await rsr.balanceOf(backingManager.address)).to.equal(0)
         expect(await rsr.balanceOf(stRSR.address)).to.equal(amount)
         expect(await rsr.balanceOf(rsrTrader.address)).to.equal(0)
@@ -687,6 +704,33 @@ describe(`Revenues - P${IMPLEMENTATION}`, () => {
         const [newRTokenTotal, newRsrTotal] = await distributor.totals()
         expect(newRsrTotal).equal(bn(60))
         expect(newRTokenTotal).equal(bn(0))
+      })
+
+      it('Should account rewards when distributing tokenToBuy', async () => {
+        // 1. StRSR.payoutRewards()
+        const stRSRBal = await rsr.balanceOf(stRSR.address)
+        await rsr.connect(owner).mint(rsrTrader.address, issueAmount)
+        await advanceTime(Number(ONE_PERIOD))
+        await expect(rsrTrader.distributeTokenToBuy()).to.emit(stRSR, 'RewardsPaid')
+        const expectedAmountStRSR = stRSRBal.add(issueAmount)
+        expect(await rsr.balanceOf(stRSR.address)).to.be.closeTo(expectedAmountStRSR, 100)
+
+        // 2. Furnace.melt()
+        // Transfer RTokens to Furnace (to trigger melting later)
+        const hndAmt: BigNumber = bn('10e18')
+        await rToken.connect(addr1).transfer(furnace.address, hndAmt)
+        await advanceTime(Number(ONE_PERIOD))
+        await furnace.melt()
+
+        // Transfer and distribute tokens in Trader (will melt)
+        await advanceTime(Number(ONE_PERIOD))
+        await rToken.connect(addr1).transfer(rTokenTrader.address, hndAmt)
+        await expect(rTokenTrader.distributeTokenToBuy()).to.emit(rToken, 'Melted')
+        const expectedAmountFurnace = hndAmt.mul(2)
+        expect(await rToken.balanceOf(furnace.address)).to.be.closeTo(
+          expectedAmountFurnace,
+          expectedAmountFurnace.div(1000)
+        ) // within 0.1%
       })
 
       it('Should update distribution even if distributeTokenToBuy() reverts', async () => {
@@ -2628,9 +2672,131 @@ describe(`Revenues - P${IMPLEMENTATION}`, () => {
           },
         ])
 
+        // Check broker disabled (batch)
+        expect(await broker.batchTradeDisabled()).to.equal(true)
+
         // Check funds at destinations
         expect(await rsr.balanceOf(stRSR.address)).to.be.closeTo(minBuyAmt.sub(10), 50)
         expect(await rToken.balanceOf(furnace.address)).to.be.closeTo(minBuyAmtRToken.sub(10), 50)
+      })
+
+      it('Should report violation even if paused or frozen', async () => {
+        // This test needs to be in this file and not Broker.test.ts because settleTrade()
+        // requires the BackingManager _actually_ started the trade
+
+        rewardAmountAAVE = bn('0.5e18')
+
+        // AAVE Rewards
+        await token2.setRewards(backingManager.address, rewardAmountAAVE)
+
+        // Collect revenue
+        // Expected values based on Prices between AAVE and RSR/RToken = 1 to 1 (for simplification)
+        const sellAmt: BigNumber = rewardAmountAAVE.mul(60).div(100) // due to f = 60%
+        const minBuyAmt: BigNumber = await toMinBuyAmt(sellAmt, fp('1'), fp('1'))
+
+        const sellAmtRToken: BigNumber = rewardAmountAAVE.sub(sellAmt) // Remainder
+        const minBuyAmtRToken: BigNumber = await toMinBuyAmt(sellAmtRToken, fp('1'), fp('1'))
+
+        // Claim rewards
+        await facadeTest.claimRewards(rToken.address)
+
+        // Check status of destinations at this point
+        expect(await rsr.balanceOf(stRSR.address)).to.equal(0)
+        expect(await rToken.balanceOf(furnace.address)).to.equal(0)
+
+        // Run auctions
+        await expectEvents(facadeTest.runAuctionsForAllTraders(rToken.address), [
+          {
+            contract: rsrTrader,
+            name: 'TradeStarted',
+            args: [anyValue, aaveToken.address, rsr.address, sellAmt, withinQuad(minBuyAmt)],
+            emitted: true,
+          },
+          {
+            contract: rTokenTrader,
+            name: 'TradeStarted',
+            args: [
+              anyValue,
+              aaveToken.address,
+              rToken.address,
+              sellAmtRToken,
+              withinQuad(minBuyAmtRToken),
+            ],
+            emitted: true,
+          },
+        ])
+
+        // Advance time till auction ended
+        await advanceTime(config.batchAuctionLength.add(100).toString())
+
+        // Perform Mock Bids for RSR and RToken (addr1 has balance)
+        // In order to force deactivation we provide an amount below minBuyAmt, this will represent for our tests an invalid behavior although in a real scenario would retrigger auction
+        // NOTE: DIFFERENT BEHAVIOR WILL BE OBSERVED ON PRODUCTION GNOSIS AUCTIONS
+        await rsr.connect(addr1).approve(gnosis.address, minBuyAmt)
+        await rToken.connect(addr1).approve(gnosis.address, minBuyAmtRToken)
+        await gnosis.placeBid(0, {
+          bidder: addr1.address,
+          sellAmount: sellAmt,
+          buyAmount: minBuyAmt.sub(10), // Forces in our mock an invalid behavior
+        })
+        await gnosis.placeBid(1, {
+          bidder: addr1.address,
+          sellAmount: sellAmtRToken,
+          buyAmount: minBuyAmtRToken.sub(10), // Forces in our mock an invalid behavior
+        })
+
+        // Freeze protocol
+        await main.connect(owner).freezeShort()
+
+        // Close auctions - Will end trades and also report violation
+        await expectEvents(facadeTest.runAuctionsForAllTraders(rToken.address), [
+          {
+            contract: broker,
+            name: 'BatchTradeDisabledSet',
+            args: [false, true],
+            emitted: true,
+          },
+          {
+            contract: rsrTrader,
+            name: 'TradeSettled',
+            args: [anyValue, aaveToken.address, rsr.address, sellAmt, minBuyAmt.sub(10)],
+            emitted: true,
+          },
+          {
+            contract: rTokenTrader,
+            name: 'TradeSettled',
+            args: [
+              anyValue,
+              aaveToken.address,
+              rToken.address,
+              sellAmtRToken,
+              minBuyAmtRToken.sub(10),
+            ],
+            emitted: true,
+          },
+          {
+            contract: rsrTrader,
+            name: 'TradeStarted',
+            emitted: false,
+          },
+          {
+            contract: rTokenTrader,
+            name: 'TradeStarted',
+            emitted: false,
+          },
+        ])
+
+        // Check broker disabled (batch)
+        expect(await broker.batchTradeDisabled()).to.equal(true)
+
+        // Funds are not distributed if paused or frozen
+        expect(await rsr.balanceOf(stRSR.address)).to.equal(0)
+        expect(await rsr.balanceOf(rsrTrader.address)).to.be.closeTo(minBuyAmt.sub(10), 50)
+        expect(await rToken.balanceOf(furnace.address)).to.equal(0)
+        expect(await rToken.balanceOf(rTokenTrader.address)).to.be.closeTo(
+          minBuyAmtRToken.sub(10),
+          50
+        )
       })
 
       it('Should not report violation when Dutch Auction clears in geometric phase', async () => {